% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinalRegressionViz-package.R
\docType{package}
\name{OrdinalRegressionViz-package}
\alias{OrdinalRegressionViz}
\alias{OrdinalRegressionViz-package}
\title{OrdinalRegressionViz: Visualize Ordinal Probit Regression Models with 'ggplot2'}
\description{
Publication-ready 'ggplot2' visualizations for ordinal probit regression models fitted with the 'ordinal' or 'brms' packages, framed in signal detection theory (SDT) terms: receiver operating characteristic (ROC) curves with uncertainty bands, latent-distribution plots with response thresholds, grouped posterior predictive checks, and forest plots for Bayesian random-effects meta-analyses. The visualizations accompany Simchon, Zipori, Teitelbaum, Lewandowsky, and van der Linden (2026) \doi{10.1016/j.copsyc.2025.102194}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tomerzipori/OrdinalRegressionViz}
  \item \url{https://tomerzipori.github.io/OrdinalRegressionViz/}
  \item Report bugs at \url{https://github.com/tomerzipori/OrdinalRegressionViz/issues}
}

}
\author{
\strong{Maintainer}: Tomer Zipori \email{tomerzip@post.bgu.ac.il}

Authors:
\itemize{
  \item Tomer Zipori \email{tomerzip@post.bgu.ac.il}
}

}
\keyword{internal}
