% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_roc_plot.R
\name{bayesian_CI_bands}
\alias{bayesian_CI_bands}
\title{Calculating credible interval bands for ROC curves}
\usage{
bayesian_CI_bands(data, x, xmin, xmax, y, ymin, ymax, group = NULL)
}
\arguments{
\item{xmin}{lower bound of CI on the x axis}

\item{xmax}{upper bound of CI on the x axis}

\item{ymin}{lower bound of CI on the y axis}

\item{ymax}{upper bound of CI on the y axis}
}
\value{
data frame
}
\description{
This is a helper function for the main Bayesian ROC curve plotting function, calculating the CI bands around the curve
}
\keyword{internal}
