% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_roc_plot.R
\name{bayesian_roc_plot}
\alias{bayesian_roc_plot}
\title{ROC Curve for two or three 2-level categorical predictor Bayesian ordinal probit regression models}
\usage{
bayesian_roc_plot(
  b_model,
  var_signal = "target",
  var_group = "time",
  var_facet = NULL,
  response = "value",
  CI = 0.95,
  centrality = "mean",
  palette = 7,
  ttl = "",
  filename = NULL,
  path = getwd(),
  width = 2450,
  height = 1446
)
}
\arguments{
\item{b_model}{Bayesian ordinal probit regression model - a brms object}

\item{var_signal}{variable analogous to the classic SDT signal, e.g. old/new}

\item{var_group}{variable name of the co-variate, e.g. experimental group}

\item{var_facet}{variable to facet by}

\item{response}{variable name of the response}

\item{CI}{credible interval width between 0 and 1}

\item{centrality}{centrality measure for posterior, either "mean" or "median"}

\item{palette}{integer representing a divergent palette in the scale_color_brewer function}

\item{ttl}{plot's title}

\item{filename}{if not NULL, the name of the png file to save the plot as}

\item{path}{folder path to save the plot in}

\item{width}{width of the saved plot - in pixels}

\item{height}{height of the saved plot - in pixels}
}
\value{
a ggplot plot object
}
\description{
This function makes a ROC curve for two or three categorical variables Bayesian ordinal probit regression models.
}
