% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_model_ppd_check.R
\name{ordinal_model_ppd_check}
\alias{ordinal_model_ppd_check}
\title{Visual posterior predictive check for Bayesian ordinal probit models. A barplot}
\usage{
ordinal_model_ppd_check(
  b_model,
  vars,
  response = "value",
  ttl = "",
  filename = NULL,
  path = getwd(),
  width = 2450,
  height = 1446
)
}
\arguments{
\item{b_model}{Bayesian ordinal probit regression model - a brms object}

\item{vars}{number of variables, either 2 or 3}

\item{response}{variable name of the response}

\item{ttl}{plot's title}

\item{filename}{if not NULL, the name of the png file to save the plot as}

\item{path}{folder path to save the plot in}

\item{width}{width of the saved plot - in pixels}

\item{height}{height of the saved plot - in pixels}
}
\value{
a ggplot plot object
}
\description{
This function creates a grouped barplot to check posterior fit to data.
}
