% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roc_ggplot_2_vars}
\alias{roc_ggplot_2_vars}
\title{ROC Curve for 2-categorical variable models}
\usage{
roc_ggplot_2_vars(
  ems_sensitivity,
  ems_specificity,
  var_signal = "target",
  var_group = "time",
  response = "value",
  palette = 2,
  ttl = ""
)
}
\arguments{
\item{ems_sensitivity}{emmeans object containing conditional probabilities for the ordinal regression model + specificity calculation for each level}

\item{var_signal}{variable analogous to the classic SDT signal, e.g. old/new}

\item{var_group}{variable name of the co-variate, e.g. experimental group}

\item{response}{variable name of the response}

\item{palette}{integer representing a divergent palette in the scale_color_brewer function}

\item{ttl}{plot's title}
}
\value{
a ggplot2::ggplot plot object
}
\description{
This function makes a ROC curve for two categorical variables ordinal regression models.
}
\keyword{internal}
