% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_roc_plot.R
\name{rvar_cumsum}
\alias{rvar_cumsum}
\title{Cumulative sum for posterior::rvar objects}
\usage{
rvar_cumsum(rvars)
}
\arguments{
\item{rvars}{a posterior rvar object}
}
\value{
numeric vector
}
\description{
This function is a simple loop to calculate a cumulative sum of rvar objects
}
\keyword{internal}
