% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_distribution_plot.R
\name{SDT_dist_ggplot}
\alias{SDT_dist_ggplot}
\title{Plot of latent normal distributions for two 2-level categorical variables ordinal probit regression models}
\usage{
SDT_dist_ggplot(
  ref_mean = 0,
  group_mean = 1,
  thresholds,
  palette = 9,
  alpha = 0.7,
  plot_limits = c(-3, 5),
  ttl = "",
  x_label = ""
)
}
\arguments{
\item{ref_mean}{mean of the normal distribution representing "noise"}

\item{group_mean}{mean of the normal distribution representing "noise + signal"}

\item{thresholds}{numeric vector representing thresholds of the ordinal model}

\item{palette}{integer representing a divergent palette in the scale_color_brewer function}

\item{alpha}{opacity of lines representing thresholds}

\item{plot_limits}{2-integer vector representing x-axis limits of the plot}

\item{ttl}{plot's title}

\item{x_label}{label of the x axis}
}
\value{
a ggplot plot object
}
\description{
This function makes a latent distributions plot for two 2-level categorical variables ordinal regression models.
}
\keyword{internal}
