% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDT_distribution_plot.R
\name{SDT_distributions_plot}
\alias{SDT_distributions_plot}
\title{Latent normal distributions for two or three 2-level categorical predictor probit ordinal regression models}
\usage{
SDT_distributions_plot(
  model,
  var_signal = "target",
  var_group = "time",
  var_facet = NULL,
  response = "value",
  plot_limits = c(-3, 5),
  palette = 9,
  alpha = 0.8,
  ttl = "",
  group1_ttl = "",
  group2_ttl = "",
  filename = NULL,
  path = getwd(),
  width = 2450,
  height = 1446
)
}
\arguments{
\item{model}{ordinal regression model, as returned by clm or clmm functions}

\item{var_signal}{variable analogous to the classic SDT signal, e.g. old/new}

\item{var_group}{variable name of the co-variate, e.g. experimental group}

\item{var_facet}{variable to facet by}

\item{response}{variable name of the response}

\item{plot_limits}{2-integer vector representing x-axis limits of the plot}

\item{palette}{integer representing a divergent palette in the scale_color_brewer function}

\item{alpha}{opacity of lines representing thresholds}

\item{ttl}{plot's title}

\item{group1_ttl}{label for the first group subplot}

\item{group2_ttl}{label for the second group subplot}

\item{filename}{if not NULL, the name of the png file to save the plot as}

\item{path}{folder path to save the plot in}

\item{width}{width of the saved plot - in pixels}

\item{height}{height of the saved plot - in pixels}
}
\value{
a ggplot plot object
}
\description{
This function makes a Latent normal distributions plot for two or three categorical variables ordinal regression models.
}
